Time: 0 ms (100.00%) | Memory: 23.3 MB (58.38%) - LeetSync

+++ 155-min-stack/min-stack.cpp
@@ -3,7 +3,7 @@ class MinStack {
     long long mini;
 public:
     MinStack() {
-        mini=LLONG_MAX;
+        mini=INT_MAX;
     }
     
     void push(int val) {
